new system constants for purge and sampling new pump set power function working sampling and purge

+++ include/Pump.h
@@ -13,7 +13,7 @@ public:
     void begin(byte _control_pin, bool _pwm);
     void begin(byte _control_pin, bool _pwm, String _name);
     void set_flow(int _flow);
-    void set_power(int _power);
+    void set_power(int8_t _power);
     int get_power();
     void start();
     void start(uint32_t time_ms);
